+++ NEWS.md
@@ -14,6 +14,8 @@
 - New `fread_colmatch()` internal for fread loading a subset of rows based on matches in a specified column
 - Add `optional` and `custom_msg` params to `package_check()`
 - Add `order` parameter to `dimFeatPlot2D` and `spatDimFeatPlot2d` to plot and order cells according to the levels of the selected feature ([#477](https://github.com/drieslab/Giotto/issues/477))
+- New `objName()` and `objName` generics for setting the names of relevant S4 subobjects
+- New `rbind()` generic to append `giottoPolygon` objects
 
 ## Changes
 - Move giotto object method-specific creation functions from `giotto.R` to `convenience.R`